+++ apps/client/src/environments/extracts-hash.ts
@@ -1,1 +1,1 @@
-export const extractsHash = `da39a3ee5e6b4b0d3255bfef95601890afd80709`;
+export const extractsHash = `cda5a15fafe82d7544882fff3e1bd69eaf9584bd`;